+++ Source/Krypton Components/Krypton.Toolkit/Controls Toolkit/KryptonCustomPaletteBase.cs
@@ -47,7 +47,7 @@ namespace Krypton.Toolkit
         private readonly PaletteRedirectCommon? _redirectCommon;
         private readonly NeedPaintHandler _needPaintDelegate;
 
-        private string _paletteFilePath;
+        private KryptonCommand _downloadPaletteDesignerCommand;
 
         #endregion
 
@@ -116,8 +116,6 @@ namespace Krypton.Toolkit
                 _basePalette.BasePaletteChanged += OnBasePaletteChanged;
                 _basePalette.BaseRendererChanged += OnBaseRendererChanged;
             }
-
-            _paletteFilePath = string.Empty;
         }
 
         /// <summary>
@@ -134,6 +132,10 @@ namespace Krypton.Toolkit
             {
                 throw new ArgumentNullException(nameof(container));
             }
+
+            _downloadPaletteDesignerCommand = new();
+
+            _downloadPaletteDesignerCommand.Execute += DownloadPaletteDesignerCommand_Execute;
 
             container.Add(this);
         }
@@ -2132,15 +2134,11 @@ namespace Krypton.Toolkit
                 if (silent)
                 {
                     ret = (string)ImportFromFile(filename);
-
-                    _paletteFilePath = filename;
                 }
                 else
                 {
                     // Perform the import operation on a separate worker thread
                     ret = CommonHelper.PerformOperation(ImportFromFile, filename) as string;
-
-                    _paletteFilePath = filename;
 
                     KryptonMessageBox.Show($"Import from file '{filename}' completed.",
                                     @"Palette Import",
@@ -3036,22 +3034,17 @@ namespace Krypton.Toolkit
                 {
                     //throw new ArgumentException($"Version '{version}' number is incompatible, only version {CURRENT_PALETTE_VERSION} or above can be imported.\nUse the PaletteUpgradeTool from the Application tab of the KryptonExplorer to upgrade.");
 
-                    DialogResult result = KryptonMessageBox.Show(
-                        $@"Unfortunately, the palette that you have chosen to import has the version: {version}, which is incompatible with {CURRENT_PALETTE_VERSION}.\nHowever, we can upgrade it now. Would you like to continue?",
-                        @"Incompatible Palette", KryptonMessageBoxButtons.YesNo, KryptonMessageBoxIcon.Question);
-
-                    if (result == DialogResult.Yes)
+                    if (version < 10)
                     {
-                        KryptonSaveFileDialog saveFileDialog = new()
-                        {
-                            Title = @"Save Krypton Palette File:",
-                            Filter = @"Krypton Palette XML Files (*.xml)|*.xml"
-                        };
 
-                        if (saveFileDialog.ShowDialog() == DialogResult.OK)
-                        {
-                            PaletteUpgradeUtility.UpgradePaletteFile(_paletteFilePath, Path.GetFullPath(saveFileDialog.FileName));
-                        }
+                    }
+                    else
+                    {
+                        DialogResult result = KryptonMessageBox.Show(
+                            $@"Unfortunately, the palette that you have chosen to import has the version: {version}, which is incompatible with {CURRENT_PALETTE_VERSION}. If you have the Palette Designer installed, you can convert it from within that application. Alternatively, you can download it from here.",
+                            @"Incompatible Palette", KryptonMessageBoxButtons.OK, KryptonMessageBoxIcon.Question,
+                            contentAreaType: MessageBoxContentAreaType.LinkLabel,
+                            linkAreaCommand: _downloadPaletteDesignerCommand, linkAreaStart: 243, linkAreaEnd: 246);
                     }
                 }
 
@@ -3687,6 +3680,19 @@ namespace Krypton.Toolkit
             }
             throw new ApplicationException($@"Unrecognised type '{s}' for import.");
         }
+
+        private void DownloadPaletteDesignerCommand_Execute(object sender, EventArgs e)
+        {
+            try
+            {
+                Process.Start(@"https://github.com/Krypton-Suite/Releases/releases");
+            }
+            catch (Exception exception)
+            {
+                ExceptionHandler.CaptureException(exception);
+            }
+        }
+
         #endregion
 
         #region Implementation GetPalette