+++ Source/Krypton Components/TestForm/Form1.Designer.cs
@@ -58,6 +58,7 @@
             this.buttonSpecAny10 = new Krypton.Toolkit.ButtonSpecAny();
             this.kryptonIntegratedToolbarPrintCommand1 = new Krypton.Toolkit.KryptonIntegratedToolbarPrintCommand();
             this.kryptonManager1 = new Krypton.Toolkit.KryptonManager(this.components);
+            this.kbtnImportPalette = new Krypton.Toolkit.KryptonButton();
             ((System.ComponentModel.ISupportInitialize)(this.kryptonPanel1)).BeginInit();
             this.kryptonPanel1.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.kryptonThemeComboBox1)).BeginInit();
@@ -65,6 +66,7 @@
             // 
             // kryptonPanel1
             // 
+            this.kryptonPanel1.Controls.Add(this.kbtnImportPalette);
             this.kryptonPanel1.Controls.Add(this.kbtnIntegratedToolbar);
             this.kryptonPanel1.Controls.Add(this.kbtnTestMessagebox);
             this.kryptonPanel1.Controls.Add(this.kryptonButton2);
@@ -75,37 +77,34 @@
             this.kryptonPanel1.Controls.Add(this.textBox1);
             this.kryptonPanel1.Dock = System.Windows.Forms.DockStyle.Fill;
             this.kryptonPanel1.Location = new System.Drawing.Point(0, 0);
-            this.kryptonPanel1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.kryptonPanel1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.kryptonPanel1.Name = "kryptonPanel1";
-            this.kryptonPanel1.Size = new System.Drawing.Size(800, 450);
+            this.kryptonPanel1.Size = new System.Drawing.Size(600, 366);
             this.kryptonPanel1.TabIndex = 0;
             // 
             // kbtnIntegratedToolbar
             // 
-            this.kbtnIntegratedToolbar.Location = new System.Drawing.Point(17, 278);
-            this.kbtnIntegratedToolbar.Margin = new System.Windows.Forms.Padding(4);
+            this.kbtnIntegratedToolbar.Location = new System.Drawing.Point(13, 226);
             this.kbtnIntegratedToolbar.Name = "kbtnIntegratedToolbar";
-            this.kbtnIntegratedToolbar.Size = new System.Drawing.Size(245, 31);
+            this.kbtnIntegratedToolbar.Size = new System.Drawing.Size(184, 25);
             this.kbtnIntegratedToolbar.TabIndex = 8;
             this.kbtnIntegratedToolbar.Values.Text = "Integrated Toolbar";
             this.kbtnIntegratedToolbar.Click += new System.EventHandler(this.kbtnIntegratedToolbar_Click);
             // 
             // kbtnTestMessagebox
             // 
-            this.kbtnTestMessagebox.Location = new System.Drawing.Point(16, 239);
-            this.kbtnTestMessagebox.Margin = new System.Windows.Forms.Padding(4);
+            this.kbtnTestMessagebox.Location = new System.Drawing.Point(12, 194);
             this.kbtnTestMessagebox.Name = "kbtnTestMessagebox";
-            this.kbtnTestMessagebox.Size = new System.Drawing.Size(247, 31);
+            this.kbtnTestMessagebox.Size = new System.Drawing.Size(185, 25);
             this.kbtnTestMessagebox.TabIndex = 7;
             this.kbtnTestMessagebox.Values.Text = "Test Messagebox";
             this.kbtnTestMessagebox.Click += new System.EventHandler(this.kbtnTestMessagebox_Click);
             // 
             // kryptonButton2
             // 
-            this.kryptonButton2.Location = new System.Drawing.Point(16, 201);
-            this.kryptonButton2.Margin = new System.Windows.Forms.Padding(4);
+            this.kryptonButton2.Location = new System.Drawing.Point(12, 163);
             this.kryptonButton2.Name = "kryptonButton2";
-            this.kryptonButton2.Size = new System.Drawing.Size(247, 31);
+            this.kryptonButton2.Size = new System.Drawing.Size(185, 25);
             this.kryptonButton2.TabIndex = 6;
             this.kryptonButton2.Values.Text = "Ribbon";
             this.kryptonButton2.Click += new System.EventHandler(this.kryptonButton2_Click);
@@ -114,10 +113,9 @@
             // 
             this.kryptonThemeComboBox1.DropDownWidth = 121;
             this.kryptonThemeComboBox1.IntegralHeight = false;
-            this.kryptonThemeComboBox1.Location = new System.Drawing.Point(16, 15);
-            this.kryptonThemeComboBox1.Margin = new System.Windows.Forms.Padding(4);
+            this.kryptonThemeComboBox1.Location = new System.Drawing.Point(12, 12);
             this.kryptonThemeComboBox1.Name = "kryptonThemeComboBox1";
-            this.kryptonThemeComboBox1.Size = new System.Drawing.Size(247, 25);
+            this.kryptonThemeComboBox1.Size = new System.Drawing.Size(185, 21);
             this.kryptonThemeComboBox1.StateCommon.ComboBox.Border.DrawBorders = ((Krypton.Toolkit.PaletteDrawBorders)((((Krypton.Toolkit.PaletteDrawBorders.Top | Krypton.Toolkit.PaletteDrawBorders.Bottom) 
             | Krypton.Toolkit.PaletteDrawBorders.Left) 
             | Krypton.Toolkit.PaletteDrawBorders.Right)));
@@ -127,11 +125,10 @@
             // kryptonButton1
             // 
             this.kryptonButton1.KryptonContextMenu = this.kryptonContextMenu1;
-            this.kryptonButton1.Location = new System.Drawing.Point(16, 162);
-            this.kryptonButton1.Margin = new System.Windows.Forms.Padding(4);
+            this.kryptonButton1.Location = new System.Drawing.Point(12, 132);
             this.kryptonButton1.Name = "kryptonButton1";
             this.kryptonButton1.ShowSplitOption = true;
-            this.kryptonButton1.Size = new System.Drawing.Size(247, 31);
+            this.kryptonButton1.Size = new System.Drawing.Size(185, 25);
             this.kryptonButton1.TabIndex = 5;
             this.kryptonButton1.Click += new System.EventHandler(this.kryptonButton1_Click);
             // 
@@ -161,18 +158,18 @@
             // 
             // kryptonListBox1
             // 
-            this.kryptonListBox1.Location = new System.Drawing.Point(269, 12);
-            this.kryptonListBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.kryptonListBox1.Location = new System.Drawing.Point(202, 10);
+            this.kryptonListBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.kryptonListBox1.Name = "kryptonListBox1";
-            this.kryptonListBox1.Size = new System.Drawing.Size(517, 426);
+            this.kryptonListBox1.Size = new System.Drawing.Size(388, 346);
             this.kryptonListBox1.TabIndex = 2;
             // 
             // kryptonTextBox1
             // 
-            this.kryptonTextBox1.Location = new System.Drawing.Point(77, 128);
-            this.kryptonTextBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.kryptonTextBox1.Location = new System.Drawing.Point(58, 104);
+            this.kryptonTextBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.kryptonTextBox1.Name = "kryptonTextBox1";
-            this.kryptonTextBox1.Size = new System.Drawing.Size(100, 27);
+            this.kryptonTextBox1.Size = new System.Drawing.Size(75, 23);
             this.kryptonTextBox1.TabIndex = 1;
             this.kryptonTextBox1.Text = "kryptonTextBox1";
             this.kryptonTextBox1.Click += new System.EventHandler(this.kryptonTextBox1_Click);
@@ -188,10 +185,10 @@
             // 
             // textBox1
             // 
-            this.textBox1.Location = new System.Drawing.Point(73, 60);
-            this.textBox1.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.textBox1.Location = new System.Drawing.Point(55, 49);
+            this.textBox1.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.textBox1.Name = "textBox1";
-            this.textBox1.Size = new System.Drawing.Size(100, 22);
+            this.textBox1.Size = new System.Drawing.Size(76, 20);
             this.textBox1.TabIndex = 0;
             this.textBox1.Click += new System.EventHandler(this.textBox1_Click);
             this.textBox1.MouseClick += new System.Windows.Forms.MouseEventHandler(this.textBox1_MouseClick);
@@ -265,9 +262,18 @@
             // 
             this.kryptonIntegratedToolbarPrintCommand1.Text = "Print";
             // 
+            // kbtnImportPalette
+            // 
+            this.kbtnImportPalette.Location = new System.Drawing.Point(13, 257);
+            this.kbtnImportPalette.Name = "kbtnImportPalette";
+            this.kbtnImportPalette.Size = new System.Drawing.Size(184, 25);
+            this.kbtnImportPalette.TabIndex = 9;
+            this.kbtnImportPalette.Values.Text = "Import Palette";
+            this.kbtnImportPalette.Click += new System.EventHandler(this.kbtnImportPalette_Click);
+            // 
             // Form1
             // 
-            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.ButtonSpecs.Add(this.buttonSpecAny1);
             this.ButtonSpecs.Add(this.buttonSpecAny2);
@@ -279,9 +285,9 @@
             this.ButtonSpecs.Add(this.buttonSpecAny8);
             this.ButtonSpecs.Add(this.buttonSpecAny9);
             this.ButtonSpecs.Add(this.buttonSpecAny10);
-            this.ClientSize = new System.Drawing.Size(800, 450);
+            this.ClientSize = new System.Drawing.Size(600, 366);
             this.Controls.Add(this.kryptonPanel1);
-            this.Margin = new System.Windows.Forms.Padding(3, 2, 3, 2);
+            this.Margin = new System.Windows.Forms.Padding(2, 2, 2, 2);
             this.Name = "Form1";
             this.PaletteMode = Krypton.Toolkit.PaletteMode.Microsoft365White;
             this.Text = "Form1";
@@ -324,5 +330,6 @@
         private Krypton.Toolkit.KryptonIntegratedToolbarPrintCommand kryptonIntegratedToolbarPrintCommand1;
         private Krypton.Toolkit.KryptonButton kbtnIntegratedToolbar;
         private Krypton.Toolkit.KryptonManager kryptonManager1;
+        private Krypton.Toolkit.KryptonButton kbtnImportPalette;
     }
 }